[none] BooksAPI.js bug
So I could not discard the changes of this file due to some weird
reasons. When I said discard the changes, it says the file doesn't
exist, while guess what, it does! I tried stash and checkout
commands with no luck. So I finally decided to just add an empty
commit and push it to the repository.