[final] Working application
This is the final version of final working application. I have not
added commited last couple of file changes since it was almost like
rewriting the entire application. I will be adding a few more
features to the app and comment the descriptions for each functions
This will be final working prototype for the app as per rubric.